:rocket: Cover removing the original data too
